ref(init): Move `sentry_sdk.init` out of `hub.py`
Now that the `Hub`-based API is deprecated, `sentry_sdk.init` should no longer be in `hub.py`. Since it is kind of its own thing, it makes sense to implement `init` in its own file.

Closes #3233
