+++ sentry_sdk/hub.py
@@ -44,7 +44,6 @@ if TYPE_CHECKING:
         LogLevelStr,
         SamplingContext,
     )
-    from sentry_sdk.consts import ClientConstructor
     from sentry_sdk.tracing import TransactionKwargs
 
     T = TypeVar("T")
@@ -57,63 +56,6 @@ else:
 
 
 _local = ContextVar("sentry_current_hub")
-
-
-class _InitGuard:
-    def __init__(self, client):
-        # type: (Client) -> None
-        self._client = client
-
-    def __enter__(self):
-        # type: () -> _InitGuard
-        return self
-
-    def __exit__(self, exc_type, exc_value, tb):
-        # type: (Any, Any, Any) -> None
-        c = self._client
-        if c is not None:
-            c.close()
-
-
-def _check_python_deprecations():
-    # type: () -> None
-    # Since we're likely to deprecate Python versions in the future, I'm keeping
-    # this handy function around. Use this to detect the Python version used and
-    # to output logger.warning()s if it's deprecated.
-    pass
-
-
-def _init(*args, **kwargs):
-    # type: (*Optional[str], **Any) -> ContextManager[Any]
-    """Initializes the SDK and optionally integrations.
-
-    This takes the same arguments as the client constructor.
-    """
-    client = Client(*args, **kwargs)
-    Scope.get_global_scope().set_client(client)
-    _check_python_deprecations()
-    rv = _InitGuard(client)
-    return rv
-
-
-from sentry_sdk._types import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    # Make mypy, PyCharm and other static analyzers think `init` is a type to
-    # have nicer autocompletion for params.
-    #
-    # Use `ClientConstructor` to define the argument types of `init` and
-    # `ContextManager[Any]` to tell static analyzers about the return type.
-
-    class init(ClientConstructor, _InitGuard):  # noqa: N801
-        pass
-
-else:
-    # Alias `init` for actual usage. Go through the lambda indirection to throw
-    # PyCharm off of the weakly typed signature (it would otherwise discover
-    # both the weakly typed signature of `_init` and our faked `init` type).
-
-    init = (lambda: _init)()
 
 
 class HubMeta(type):
